Changed handling of $INCLUDE files:
Files included with $INCLUDE are not relative to the zone file
itself, but to the options->directory statement in named.conf.

This makes $INCLUDE statements using relative paths work for
all zone files. Until now, this was broken for zone files that
were located in a subdirectory. (For such files, dependencies
were not computed correctly.